requirements.txt
matplotlib==3.8.4
numpy==2.1.1
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==10.4.0
preprocessing==0.1.13
scikit_learn==1.4.2
scipy==1.14.1
seaborn==0.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.4
transformers==4.45.1
